[wlan][iwlwifi] Delete the RX allocator
The existing driver implements zero-copy receive by unmapping the DMA
buffer from the device and passing it up the stack. This means that the
driver needs to continuously allocate more buffers to replace the ones
it has lost.

In Fuchsia the IOMMU will be enabled by default, which makes pinning and
unpinning VMOs an expensive operation. It will be more efficient to
simply copy the data out of the VMOs instead of unpin and reallocate.
Therefore, we no longer need the allocator code.

Bug:35409
TEST=fx run-test iwlwifi_tests. Behaviour should be the same.

Change-Id: Ib94ee271edbdfc4a547e2e3cb37e56252fbc67bc