[zircon] Implement a generic fake-object for defining specialized fakes
The object and handle table support in fake-bti has been pulled out and
expanded to provide the fake-object library. Additionally, object has
been extended to support implementations for all of the zx_object_*
syscalls so they can be built for fakes, or passed through in the event
real handles are passed to them. This provides a scaffolding to build
other fake objects based on the work done for fake-bti.

Specialized creation methods may still be needed for some objects that
are privileged (like a bti or fake root resource), but the dtors for
zx:: objects and zx_handle_close() can properly close/clean up these
fake objects now. Due to this fake_bti_destroy has been removed and the
expectation is no future objects should need to implement a similar api.

The removal of fake_bti_destroy required modifying several tests to move
to the newer zx_handle_close() usage. In general, if the code is C++ it
was migrated to use zx::bti, and if it was C then zx_handle_close is now
used.

This CL is a re-re-landed version with a soft api transition now that we
know CQ doesn't run the tests I thought they ran. GI lifetime issues
have now been fixed so this will be the third in the Fake Object
Trilogy.

Change-Id: Iffbf769811a023b5edec937c450d041c8cd21b67