[sim-env][sim-fake-ap] Merge Rx*() and Tx*() into Rx() and Tx()
Introduce Frame classes to be passed as parameter of Rx() and Tx()
function, so that we don't need all the Rx*() and Tx*() functions,
to deal with different kinds of packets transmission anymore.

Receive handlers in brcmfmac's sim_hw are also combined into one, which leave
the frame classification work to sim_fw.

Bug: 43702
Test: wlan-sim-env-test/wlan-sim-fake-ap-test/brcmfmac_sim_test

Change-Id: I86b3356757367cafcabaf41fe4a1178acf2fdc9d