[wlan][iwlwifi] Move io_buffer_t functionality to iwl_iobuf
Fuchsia uses io_buffer_t instances to allocate and manage DMA buffers.
As this is a Fuchsia-specific mechanism, we introduce here an abstracted
iwl_iobuf struct, and implement it with an io_buffer_t under the hood
for Fuchsia.

This in particular also allows us to remove the direct inclusion of
<lib/ddk/io_buffer.h> from many C headers that pose a compilation
problem in C++ mode.

Bug: 75293, 76234
Test: fx test iwlwifi_test, WLAN scan on a 7265 chipset on NUC

Change-Id: If87466027a7e5ca6422f23c16c49673270f83ccc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535263
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Reviewed-by: John Sheu <sheu@google.com>
Commit-Queue: John Sheu <sheu@google.com>