[wlan][iwlwifi] Ensure alignment of SimMvm resp_buf_
SimMvm contains a member of type SimNvm, which has zero members and thus
is a 1-byte sized member.  The next member of SimMvm is resp_buf_, used
as a response data buffer and declared as a uint8_t array, which has an
alignment requirement of 1 byte (as its elements have 1-byte alignment);
thus the array's address starts at a 1 byte offset.

Later on, resp_buf_ is cast into other struct types, but as it has a 1
byte offset from a typical aligned value, this causes undefined behavior
(and UBSan errors) down the line when multi-byte struct members are read
or written in an unaligned manner from the aliased struct.

In this commit we move to allocating resp_buf_ dynamically, which
ensures that the buffer is aligned to at least the platform's strictest
fundamental alignment requirement.

Bug: 46688
Test: fx test iwlwifi_test
Change-Id: I1cc05c3c5b05bace2f6681858b1de8cbb073a427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544601
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Commit-Queue: John Sheu <sheu@google.com>