[wlan][iwlwifi] port mac_clear_assoc()
This is the last piece of code that enables the disassociation. The
changes include the Tx queue management and the STA state transition.

This is also the follow-up of fxr/56379. The mvm_sta ownership is
actually transferred into mvm->fw_id_to_mac_id[] while configure_bss.
So we cannot remove it while we are still using it.

Bug: 36742, 49531
Change-Id: I4a265ba96ee276e995e4589e319ef3b8b8084705
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412252
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>