[wlan][iwlwifi] Remove unnecessary assertion for associaiton.
The original code (for 7265) has incorrect assumption that
the driver cannot support the chipsets with low latency mode.
It turns out the assumption is overthinking.

We are still able to associate even if the chipset supports
it but doesn't enable it. So remove the assertion.

Bug: 51985
Tested: Verified on NUC 8265. Associated with an open network.
Change-Id: I7676509f91b4278ca01408d48ffc87d0a03b95af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562045
Reviewed-by: Joshua Conner <joshconner@google.com>
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>