[wlan][iwlwifi] unbind MAC context in mac_set_channel (if bound)
If an association failed (e.g. AP rejected the client), the driver
could leave the firmware in the some state that crashes the next
association. This is why we saw BINDING_CONTEXT_CMD crash, which
the binding had been set in the firmware.

Since there is no mechanism for MLME to inform the driver that
the association is failed, we check this in the driver. If the
info has been set already, we clean it up before adding new one.

Bug: 86305

Test: Added a test to check if set_channel handles being called
before clear_assoc() is called.

Change-Id: I104227d2c78d3a7d2e1e603133c15c6d77de2d5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591707
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>