[iwlwifi] Support HT initial rate setting
The goal of this change is to set an initial 802.11n rate to the
firmware when association(to an AP which supports HT) is done.

The major part of this porting effort is the rate scale logic in
rateScaleMng.c, the rate selection part is not ported in this change.
Also some auxiliary functions are ported to make this effort more
smooth.

Bug: 86522, 89011
Test: New test case added in wlan-softmac-device-test.cc.

Change-Id: I44db126f79ecd85002f904f06fc5d8bd6924628e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585781
Reviewed-by: Yung-chieh Lo <yjlou@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>