[iwlwifi] remove the duplicate command
The first iwl_mvm_configure_filter() call is no-op because
the mvmvif->bss_conf.assoc is still false (checked in
iwl_mvm_mc_iface_iterator()).

This is the code left unexpectedly during a refactoring.
Remove it in order to remove the error message in the log.

Tested: The unittest code is unchanged because the first call is no-op.
Change-Id: I60406158d7185e1acb521fbae692efe233ea233d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642233
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Yung-chieh Lo <yjlou@google.com>