[iwlwifi] port iwl_parse_nvm_mcc_info()
This code parses the regulatory response from the firmware
and store the data in mvm->mcc_info. The key info is the
allowed channel list for the current country.

The CL also contains the logic how we filter the channel list
from the upper layer.

Bug: 87321
Tested: Tested on the device. Added test cases.
Change-Id: Ie512833b1a414f170b7e3d12ccf781f60b98d810
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647587
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Yung-chieh Lo <yjlou@google.com>